+++ src/register/apsr.rs
@@ -1,5 +1,8 @@
 //! Application Program Status Register
 
+#[cfg(cortex_m)]
+use core::arch::asm;
+
 /// Application Program Status Register
 #[derive(Clone, Copy, Debug)]
 pub struct Apsr {
@@ -45,10 +48,10 @@ impl Apsr {
 }
 
 /// Reads the CPU register
-///
-/// **NOTE** This function is available if `cortex-m` is built with the `"inline-asm"` feature.
+#[cfg(cortex_m)]
 #[inline]
 pub fn read() -> Apsr {
-    let bits: u32 = call_asm!(__apsr_r() -> u32);
+    let bits;
+    unsafe { asm!("mrs {}, APSR", out(reg) bits, options(nomem, nostack, preserves_flags)) };
     Apsr { bits }
 }