+++ src/register/primask.rs
@@ -1,5 +1,8 @@
 //! Priority mask register
 
+#[cfg(cortex_m)]
+use core::arch::asm;
+
 /// All exceptions with configurable priority are ...
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Primask {
@@ -24,9 +27,11 @@ impl Primask {
 }
 
 /// Reads the CPU register
+#[cfg(cortex_m)]
 #[inline]
 pub fn read() -> Primask {
-    let r: u32 = call_asm!(__primask_r() -> u32);
+    let r: u32;
+    unsafe { asm!("mrs {}, PRIMASK", out(reg) r, options(nomem, nostack, preserves_flags)) };
     if r & (1 << 0) == (1 << 0) {
         Primask::Inactive
     } else {